Refactor: rename func
